refactor(ast): modularize ast form and step1
