+++ components/ASTForm.tsx
@@ -1,16 +1,16 @@
 'use client';
 
 import React, { useState, useEffect, useCallback, useRef, useMemo } from 'react';
-import { 
-  FileText, ArrowLeft, ArrowRight, Save, Eye, Download, CheckCircle, 
-  AlertTriangle, Clock, Shield, Users, MapPin, Calendar, Building, 
+import {
+  FileText, ArrowLeft, ArrowRight, Save, Eye, Download, CheckCircle,
+  AlertTriangle, Clock, Shield, Users, MapPin, Calendar, Building,
   Phone, User, Briefcase, Copy, Check, Camera, HardHat, Zap, Settings,
   Plus, Trash2, Edit, Star, Wifi, WifiOff, Upload, Bell, Wrench, Wind,
   Droplets, Flame, Activity, Search, Filter, Hand, MessageSquare
 } from 'lucide-react';
 
 // =================== ✅ IMPORTS DES COMPOSANTS STEPS 1-6 (CONSERVÉS INTÉGRALEMENT) ===================
-import Step1ProjectInfo from '@/components/steps/Step1ProjectInfo';
+import { Step1ProjectInfo, steps, translations, useIsMobile } from '@/components/ast';
 import Step2Equipment from '@/components/steps/Step2Equipment';
 import Step3Hazards from '@/components/steps/Step3Hazards';
 import Step4Permits from '@/components/steps/Step4Permits';
@@ -26,203 +26,6 @@ interface ASTFormProps {
   formData: any;
   onDataChange: (section: string, data: any) => void;
 }
-
-// =================== TRADUCTIONS COMPLÈTES (CONSERVÉES INTÉGRALEMENT) ===================
-const translations = {
-  fr: {
-    title: "🛡️ C-Secur360",
-    subtitle: "Analyse Sécuritaire de Travail",
-    systemOperational: "Système opérationnel",
-    astStep: "AST • Étape",
-    astNumber: "NUMÉRO AST",
-    online: "En ligne",
-    offline: "Hors ligne",
-    submit: "Soumettre",
-    approve: "Approuver",
-    status: {
-      draft: "Brouillon",
-      pending_verification: "En attente",
-      approved: "Approuvé",
-      auto_approved: "Auto-approuvé",
-      rejected: "Rejeté"
-    },
-    progress: "Progression AST",
-    completed: "complété",
-    stepOf: "sur",
-    previous: "Précédent",
-    next: "Suivant",
-    finished: "Terminé ✓",
-    autoSave: "Sauvegarde auto",
-    saving: "Modification...",
-    saved: "Sauvegardé",
-    active: "Actif",
-    language: "Langue",
-    french: "Français",
-    english: "English",
-    steps: {
-      step1: {
-        title: "Informations Projet",
-        subtitle: "Identification & Verrouillage"
-      },
-      step2: {
-        title: "Équipements", 
-        subtitle: "EPI et équipements sécurité"
-      },
-      step3: {
-        title: "Dangers & Contrôles",
-        subtitle: "Risques + Moyens contrôle"
-      },
-      step4: {
-        title: "Permis & Autorisations",
-        subtitle: "Conformité réglementaire"
-      },
-      step5: {
-        title: "Validation Équipe",
-        subtitle: "Signatures & Approbations"
-      },
-      step6: {
-        title: "Finalisation",
-        subtitle: "Consentement & Archive"
-      }
-    }
-  },
-  en: {
-    title: "🛡️ C-Secur360",
-    subtitle: "Job Safety Analysis",
-    systemOperational: "System operational",
-    astStep: "JSA • Step",
-    astNumber: "JSA NUMBER",
-    online: "Online",
-    offline: "Offline",
-    submit: "Submit",
-    approve: "Approve",
-    status: {
-      draft: "Draft",
-      pending_verification: "Pending",
-      approved: "Approved",
-      auto_approved: "Auto-approved",
-      rejected: "Rejected"
-    },
-    progress: "JSA Progress",
-    completed: "completed",
-    stepOf: "of",
-    previous: "Previous",
-    next: "Next",
-    finished: "Finished ✓",
-    autoSave: "Auto save",
-    saving: "Saving...",
-    saved: "Saved",
-    active: "Active",
-    language: "Language",
-    french: "Français",
-    english: "English",
-    steps: {
-      step1: {
-        title: "Project Information",
-        subtitle: "Identification & Lockout"
-      },
-      step2: {
-        title: "Equipment",
-        subtitle: "PPE and safety equipment"
-      },
-      step3: {
-        title: "Hazards & Controls",
-        subtitle: "Risks + Control measures"
-      },
-      step4: {
-        title: "Permits & Authorizations",
-        subtitle: "Regulatory compliance"
-      },
-      step5: {
-        title: "Team Validation",
-        subtitle: "Signatures & Approvals"
-      },
-      step6: {
-        title: "Finalization",
-        subtitle: "Consent & Archive"
-      }
-    }
-  }
-};
-
-// =================== CONFIGURATION DES STEPS (CONSERVÉE) ===================
-const steps = [
-  {
-    id: 1,
-    titleKey: 'step1',
-    icon: FileText,
-    color: '#3b82f6',
-    required: true
-  },
-  {
-    id: 2,
-    titleKey: 'step2',
-    icon: Shield,
-    color: '#10b981',
-    required: true
-  },
-  {
-    id: 3,
-    titleKey: 'step3',
-    icon: AlertTriangle,
-    color: '#f59e0b',
-    required: true
-  },
-  {
-    id: 4,
-    titleKey: 'step4',
-    icon: Edit,
-    color: '#8b5cf6',
-    required: false
-  },
-  {
-    id: 5,
-    titleKey: 'step5',
-    icon: Users,
-    color: '#06b6d4',
-    required: false
-  },
-  {
-    id: 6,
-    titleKey: 'step6',
-    icon: CheckCircle,
-    color: '#10b981',
-    required: false
-  }
-];
-
-// =================== HOOK DÉTECTION MOBILE OPTIMISÉ (CONSERVÉ) ===================
-const useIsMobile = () => {
-  const [isMobile, setIsMobile] = useState(() => {
-    if (typeof window !== 'undefined') {
-      return window.innerWidth <= 768;
-    }
-    return false;
-  });
-
-  useEffect(() => {
-    let timeoutId: NodeJS.Timeout;
-    
-    const checkIsMobile = () => {
-      clearTimeout(timeoutId);
-      timeoutId = setTimeout(() => {
-        const newIsMobile = window.innerWidth <= 768;
-        if (newIsMobile !== isMobile) {
-          setIsMobile(newIsMobile);
-        }
-      }, 150);
-    };
-
-    window.addEventListener('resize', checkIsMobile);
-    
-    return () => {
-      window.removeEventListener('resize', checkIsMobile);
-      clearTimeout(timeoutId);
-    };
-  }, [isMobile]);
-
-  return isMobile;
-};
 
 // =================== COMPOSANT PRINCIPAL ASTFORM ===================
 export default function ASTForm({
