+++ components/ast/Step1ProjectInfo/__tests__/helpers.test.ts
@@ -1,0 +1,10 @@
+import { generateASTNumber } from '../helpers';
+
+describe('generateASTNumber', () => {
+  it('generates unique numbers', () => {
+    const a = generateASTNumber();
+    const b = generateASTNumber();
+    expect(a).not.toBe(b);
+    expect(a).toMatch(/^AST-/);
+  });
+});
