+++ components/ast/Step1ProjectInfo/helpers.ts
@@ -1,0 +1,8 @@
+export const generateASTNumber = (): string => {
+  const year = new Date().getFullYear();
+  const month = String(new Date().getMonth() + 1).padStart(2, '0');
+  const day = String(new Date().getDate()).padStart(2, '0');
+  const timestamp = Date.now().toString().slice(-6);
+  const random = Math.floor(Math.random() * 9999).toString().padStart(4, '0');
+  return `AST-${year}${month}${day}-${timestamp}${random.slice(0, 2)}`;
+};
