+++ components/ast/Step1ProjectInfo/types.ts
@@ -1,0 +1,58 @@
+export interface WorkLocation {
+  id: string;
+  name: string;
+  description: string;
+  zone: string;
+  building?: string;
+  floor?: string;
+  maxWorkersReached: number;
+  currentWorkers: number;
+  lockoutPoints: number;
+  isActive: boolean;
+  createdAt: string;
+  notes?: string;
+  estimatedDuration: string;
+  startTime?: string;
+  endTime?: string;
+}
+
+export interface LockoutPoint {
+  id: string;
+  energyType: 'electrical' | 'mechanical' | 'hydraulic' | 'pneumatic' | 'chemical' | 'thermal' | 'gravity';
+  equipmentName: string;
+  location: string;
+  lockType: string;
+  tagNumber: string;
+  isLocked: boolean;
+  verifiedBy: string;
+  verificationTime: string;
+  photos: string[];
+  notes: string;
+  completedProcedures: number[];
+  assignedLocation?: string;
+}
+
+export interface LockoutPhoto {
+  id: string;
+  url: string;
+  caption: string;
+  category: 'before_lockout' | 'during_lockout' | 'lockout_device' | 'client_form' | 'verification';
+  timestamp: string;
+  lockoutPointId?: string;
+}
+
+export interface LocationStats {
+  totalWorkers: number;
+  totalLocations: number;
+  activeLockouts: number;
+  peakUtilization: number;
+  locationBreakdown: {
+    locationId: string;
+    name: string;
+    currentWorkers: number;
+    maxReached: number;
+    lockouts: number;
+    utilizationCurrent: number;
+    estimatedDuration: string;
+  }[];
+}
