+++ components/ast/__tests__/stepsConfig.test.ts
@@ -1,0 +1,7 @@
+import { steps } from '../stepsConfig';
+
+describe('stepsConfig', () => {
+  it('defines six steps', () => {
+    expect(steps).toHaveLength(6);
+  });
+});
