+++ components/ast/__tests__/useIsMobile.test.ts
@@ -1,0 +1,10 @@
+import { renderHook } from '@testing-library/react';
+import { useIsMobile } from '../useIsMobile';
+
+describe('useIsMobile', () => {
+  it('detects mobile width', () => {
+    Object.defineProperty(window, 'innerWidth', { writable: true, configurable: true, value: 500 });
+    const { result } = renderHook(() => useIsMobile());
+    expect(result.current).toBe(true);
+  });
+});
