+++ components/ast/index.ts
@@ -1,0 +1,4 @@
+export { translations } from './translations';
+export { steps } from './stepsConfig';
+export { useIsMobile } from './useIsMobile';
+export { default as Step1ProjectInfo } from './Step1ProjectInfo';
