+++ components/ast/useIsMobile.ts
@@ -1,0 +1,33 @@
+import { useEffect, useState } from 'react';
+
+export const useIsMobile = () => {
+  const [isMobile, setIsMobile] = useState(() => {
+    if (typeof window !== 'undefined') {
+      return window.innerWidth <= 768;
+    }
+    return false;
+  });
+
+  useEffect(() => {
+    let timeoutId: NodeJS.Timeout;
+
+    const checkIsMobile = () => {
+      clearTimeout(timeoutId);
+      timeoutId = setTimeout(() => {
+        const newIsMobile = window.innerWidth <= 768;
+        if (newIsMobile !== isMobile) {
+          setIsMobile(newIsMobile);
+        }
+      }, 150);
+    };
+
+    window.addEventListener('resize', checkIsMobile);
+
+    return () => {
+      window.removeEventListener('resize', checkIsMobile);
+      clearTimeout(timeoutId);
+    };
+  }, [isMobile]);
+
+  return isMobile;
+};
