+++ package.json
@@ -9,7 +9,8 @@
     "lint": "next lint",
     "db:generate": "prisma generate",
     "db:push": "prisma db push",
-    "postinstall": "prisma generate"
+    "postinstall": "prisma generate",
+    "test": "vitest run"
   },
   "dependencies": {
     "react": "^18",
@@ -43,6 +44,8 @@
     "@types/bcryptjs": "^2.4.6",
     "eslint": "^8",
     "eslint-config-next": "14.0.4",
-    "@types/qrcode": "^1.5.5"
+    "@types/qrcode": "^1.5.5",
+    "vitest": "^0.34.0",
+    "@testing-library/react": "^14.0.0"
   }
 }
